+++ plugins/main/public/components/common/data-grid/use-data-grid.tsx
@@ -17,6 +17,7 @@ import {
   IndexPattern,
 } from '../../../../../../src/plugins/data/common';
 import { EuiDataGridPaginationProps } from '@opensearch-project/oui';
+import dompurify from 'dompurify';
 
 export interface PaginationOptions
   extends Pick<
@@ -163,7 +164,21 @@ export const useDataGrid = (props: tDataGridProps): EuiDataGridProps => {
         );
       }
 
-      return fieldFormatted;
+      // Format the value using the field formatter
+      // https://github.com/opensearch-project/OpenSearch-Dashboards/blob/2.16.0/src/plugins/discover/public/application/components/data_grid/data_grid_table_cell_value.tsx#L80-L89
+      const formattedValue = indexPattern.formatField(
+        rows[relativeRowIndex],
+        columnId,
+      );
+      if (typeof formattedValue === 'undefined') {
+        return <span>-</span>;
+      } else {
+        const sanitizedCellValue = dompurify.sanitize(formattedValue);
+        return (
+          // eslint-disable-next-line react/no-danger
+          <span dangerouslySetInnerHTML={{ __html: sanitizedCellValue }} />
+        );
+      }
     }
     return null;
   };