+++ plugins/main/public/components/common/data-source/pattern/pattern-data-source.ts
@@ -11,6 +11,7 @@ import {
 } from '../../../../../../../src/plugins/data/public';
 import { search } from '../../search-bar/search-bar-service';
 import { PatternDataSourceFilterManager } from './pattern-data-source-filter-manager';
+import { useSelector } from 'react-redux';
 
 export class PatternDataSource implements tDataSource {
   id: string;
@@ -40,28 +41,27 @@ export class PatternDataSource implements tDataSource {
   }
 
   getFetchFilters(): tFilter[] {
-    return [...this.getAllowAgentsFilter(), ...this.getExcludeManagerFilter()];
+    return [...this.getExcludeManagerFilter()];
   }
 
   async select() {
-    let pattern: IndexPattern;
     try {
-      pattern = await this.patternService.get(this.id);
-      if (!pattern)
+      const pattern = await this.patternService.get(this.id);
+      if (pattern) {
+        const fields = await this.patternService.getFieldsForIndexPattern(
+          pattern,
+        );
+        const scripted = pattern.getScriptedFields().map(field => field.spec);
+        pattern.fields.replaceAll([...fields, ...scripted]);
+        try {
+          await this.patternService.updateSavedObject(pattern);
+        } catch {}
+      } else {
         throw new Error('Error selecting index pattern: pattern not found');
-
-      const fields = await this.patternService.getFieldsForIndexPattern(
-        pattern,
-      );
-      const scripted = pattern.getScriptedFields().map(field => field.spec);
-      pattern.fields.replaceAll([...fields, ...scripted]);
+      }
     } catch (error) {
       throw new Error(`Error selecting index pattern: ${error}`);
     }
-    try {
-      // Vulnerability dashboard error loading for read only user
-      await this.patternService.updateSavedObject(pattern);
-    } catch {}
   }
 
   async fetch(params: tSearchParams) {
@@ -123,10 +123,6 @@ export class PatternDataSource implements tDataSource {
    */
   getPinnedAgentFilter(): tFilter[] {
     return PatternDataSourceFilterManager.getPinnedAgentFilter(this.id);
-  }
-
-  getAllowAgentsFilter(): tFilter[] {
-    return PatternDataSourceFilterManager.getAllowAgentsFilter(this.id);
   }
 
   getExcludeManagerFilter(): tFilter[] {