+++ plugins/main/server/lib/reporting/base-query.ts
@@ -12,16 +12,16 @@
 
 import { cloneDeep } from 'lodash';
 
-export function Base(pattern: string, filters: any, gte: number, lte: number, allowedAgentsFilter: any = null) {
+export function Base(pattern: string, filters: any, gte: number, lte: number) {
   const clonedFilter = cloneDeep(filters);
   clonedFilter?.bool?.must?.push?.({
     range: {
       timestamp: {
         gte: gte,
         lte: lte,
-        format: 'epoch_millis'
-      }
-    }
+        format: 'epoch_millis',
+      },
+    },
   });
   const base = {
     from: 0,
@@ -29,7 +29,7 @@ export function Base(pattern: string, filters: any, gte: number, lte: number, al
     aggs: {},
     sort: [],
     script_fields: {},
-    query: clonedFilter
+    query: clonedFilter,
   };
 
   return base;